Implement CSR
When using the traditional HTML anchor element, when a link is clicked a
full page reload happens.

Next.js provides the Link component which allows you to implement Client
Side Rendering.

The advantage over a traditional React SPA is that Next.js code split
your application by routes. This means that, while in a traditional SPA
all the UI will be loaded in the first request (if you haven't code
split the app manually), a Next.js app will load only the UI of the page
users navigate to.

Furthermore, in production, whenever Link components appear in the
browser's viewport, Next.js automatically prefetches the code for the
linked route in the background. By the time the user clicks, the link,
the code for the destination page will already be loaded in the
background, and this is what makes the page transition near-instant!

The only thing you need to do is to replace `<a>` by `<Link />` in your
code.

If you want to show the active link, you can get the current path using
the `usePathname()` hook and conditionally style the link using the clsx
library.

`const pathname = usePathname()` and `pathname === link.href` are vital
for implementing the technique.