Implement search functionality
In a traditional setup, you would save the searched term in state.
However, using URL search params to manage the search state gives you
these benefits:

 - Bookmarkable and Sharareable URLs

 - Server-Side Rendering and Initial Load

 - Analytics and Tracking

The search functionality spans the client and the server. When a user
searches for an invoice on the client, the URL params will be updated,
data will be fetched on the server, and the table will re-render on the
server with the new data.

To add the search query to the URL, we

 - Listen for changes on the input field and call a handler function to
   update the URL

 - Inside the handler, create an instance of URLSearchParams() to access
   utility methods to work with the query string

 - Check if the user typed something and if so, set the `query` key
   equals to the inserted value. If not, remove the `query` key from the
   query string. We achieve these using utility methods

 - Then, use the `replace()` function of Next.js's `useRouter()` hook to
   update the URL by appending the query string to it

 - Finally, keep the value of the search input field in sync with the
   URL (Changing the value assigned to the `query` key directly in the
   URL and pressing enter updates the value of the input field)

The last step uses the `defaultValue` prop instead of `value` because,
if you're using state to manage the value of an input, you'd use the
`value` attribute to make it a controlled component. This means React
would manage the input's state. However, since you're not using state,
you can use `defaultValue`. This means the native input will manage its
own state. This is okay since you're saving the search query to the URL
instead of state.