Display search results
The fetching and rendering of data happens in a server component.

You cannot use hooks in server components but, to render the results,
the Page component needs to access both `query` and `currentPage`. The
solution is to destructure `props` and access `searchParams`.

As a general rule, if you want to read the params from the client, use
the `useSearchParams()` hook. If you want to read params from the
server, use the `searchParams` prop.