Implement the create invoices functionality
In HTML, the form data is submitted via the action attribute to an URL
for processing. In React, we can submit the data, the native FormData to
a Server Action instead.

You just need to specify the action attribute and, instead of passing an
URL to it, pass a Server Action.

With Server Components, you don't need an API endpoint for handling
database manipulation. You can do it directly from the action.

Here we're creating an `actions.ts` file, but you can also write Server
Actions directly inside Server Components by adding "use server"
_inside_ the action.

Inside the action, you do the usuals:

 - Extract FormData

 - Validate the received data

 - Prepare the data (form example, here we convert the amount in cents
   before storing)

 - Store the data

 - Clear client cache since we want to load the new page sent by the
   server

 - Redirect the user to the updated page

There are several ways to perform validation. In this project, the Zod
library is being used.