Handle all errors with error.tsx
An error file defines an error UI boundary for a route segment. It is
useful for catching unexpected errors that occur in Server Components
and Client Components and displaying a fallback UI.

The Error component receives two props:

 - `error`: An instance of JavaScript's native Error object

 - `reset`: A function to reset the error boundary. When executed, the
   function will try to re-render the Error boundary's contents. If
successful, the fallback error component is replaced with the result of
the re-render

The Error component is rendered for this route and for its children.